V0.8.0

+++ src/librabbitmq/amqp_api.c
@@ -267,7 +267,7 @@ amqp_rpc_reply_t amqp_channel_close(amqp_connection_state_t state,
   amqp_method_number_t replies[2] = { AMQP_CHANNEL_CLOSE_OK_METHOD, 0};
   amqp_channel_close_t req;
 
-  req.reply_code = code;
+  req.reply_code = (uint16_t)code;
   req.reply_text.bytes = codestr;
   req.reply_text.len = sprintf(codestr, "%d", code);
   req.class_id = 0;
@@ -284,7 +284,7 @@ amqp_rpc_reply_t amqp_connection_close(amqp_connection_state_t state,
   amqp_method_number_t replies[2] = { AMQP_CONNECTION_CLOSE_OK_METHOD, 0};
   amqp_channel_close_t req;
 
-  req.reply_code = code;
+  req.reply_code = (uint16_t)code;
   req.reply_text.bytes = codestr;
   req.reply_text.len = sprintf(codestr, "%d", code);
   req.class_id = 0;
